server/Azure_FunctionApp/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
mysql-connector-python==8.1.0
opencv-python-headless==4.8.1.78
numpy==1.26.1
Pillow==10.1.0
requests==2.31.0